+++ src/Symfony/Bundle/FrameworkBundle/Resources/config/console.php
@@ -60,6 +60,7 @@ use Symfony\Component\Translation\Command\TranslationLintCommand;
 use Symfony\Component\Translation\Command\TranslationPullCommand;
 use Symfony\Component\Translation\Command\TranslationPushCommand;
 use Symfony\Component\Translation\Command\XliffLintCommand;
+use Symfony\Component\Translation\Command\XliffUpdateSourcesCommand;
 use Symfony\Component\Validator\Command\DebugCommand as ValidatorDebugCommand;
 use Symfony\Component\Workflow\Command\WorkflowDumpCommand;
 use Symfony\Component\Yaml\Schema\FileHeaderSchemaResolver;
@@ -324,6 +325,16 @@ return static function (ContainerConfigurator $container) {
             ->tag('console.command')
 
         ->set('console.command.xliff_lint', XliffLintCommand::class)
+            ->tag('console.command')
+
+        ->set('console.command.translation_xliff_update_sources', XliffUpdateSourcesCommand::class)
+            ->args([
+                service('translation.writer'),
+                service('translation.reader'),
+                param('kernel.default_locale'),
+                [], // Translator paths
+                param('kernel.enabled_locales'),
+            ])
             ->tag('console.command')
 
         ->set('console.command.yaml_lint', YamlLintCommand::class)